refactor: file restructuring and moving files to appropriate folders, added new dependency for components (shadCN)
