+++ package.json
@@ -18,6 +18,9 @@
         "@tailwindcss/vite": "^4.0.6",
         "axios": "^1.8.1",
         "axios-hooks": "^5.1.0",
+        "class-variance-authority": "^0.7.1",
+        "clsx": "^2.1.1",
+        "lucide-react": "^0.479.0",
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
         "react-hook-form": "^7.54.2",
@@ -25,7 +28,9 @@
         "react-redux": "^9.2.0",
         "react-router": "^7.2.0",
         "react-router-dom": "^6.30.0",
+        "tailwind-merge": "^3.0.2",
         "tailwindcss": "^4.0.6",
+        "tailwindcss-animate": "^1.0.7",
         "zod": "^3.24.2"
     },
     "devDependencies": {
