+++ src/components/shared/Input.tsx
@@ -26,7 +26,7 @@ export default function Input({
                     {...(!label && { 'aria-label': placeholder })}
                     placeholder={type === 'password' ? '********' : placeholder}
                     type={type === 'password' ? (showPassword ? 'text' : 'password') : type}
-                    className={`w-full h-full rounded-lg border-2 border-black/25 py-4 px-5 font-inter placeholder:font-normal placeholder:text-black/30 focus:outline-none focus:ring-2 focus:border-0 focus:ring-primary`}
+                    className={`w-full h-full rounded-lg border-2 border-black/25 py-4 px-5 font-inter placeholder:font-normal placeholder:text-black/30 focus:outline-none focus:ring-2 focus:border-0 focus:ring-primary-600`}
                 />
                 {error && <p className="py-1 font-Inter text-sm text-red-400">{error.message}</p>}
                 {type === 'password' && (
