+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export const API_URL = import.meta.env.VITE_API_URL
