+++ src/pages/landing.tsx
@@ -7,22 +7,21 @@ export default function LandingPage() {
         <div className="w-full flex flex-col px-4 py-2 md:px-24 md:py-12 h-screen">
             <FocusFlowHeader />
             <div className="flex justify-between mt-20 items-center">
-                <div className="flex lg:flex-col gap-12 flex-col text-center items-center">
-                    <p className="text-5xl font-bold font-inter text-secondary ">
-                        Where focus meets <br /> productivity
+                <div className="w-1/2 flex lg:flex-col gap-12 flex-col text-left">
+                    <p className="text-5xl font-bold font-inter text-primary-700">
+                        Where focus meets productivity
                     </p>
-                    <p className="font-inter text-xl">
-                        Streamlining your day and <br />
-                        making every moment count.
+                    <p className="font-inter text-3xl w-2/3">
+                        Streamlining your day and making every moment count.
                     </p>
                     <Link
-                        className="w-48 bg-primary shadow shadow-black/40 drop-shadow-xl text-white rounded-lg text-center p-4"
+                        className="w-52 bg-primary-600 shadow text-xl shadow-black/40 drop-shadow-xl text-white rounded-lg text-center p-4"
                         to="/sign-up"
                     >
                         Get started
                     </Link>
                 </div>
-                <img src={LandingImage} alt="landing page Image" className="lg:block hidden" />
+                <img src={LandingImage} alt="landing page Image" className="lg:block hidden w-2/5 mr-20" />
             </div>
         </div>
     )
