+++ src/routes/index.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from 'react-router-dom'
 import LandingPage from '../pages/landing'
 import SignUpPage from '../pages/sign-up'
-import LoginPage from '../pages/log-in'
+import LoginPage from '../pages/login'
 
 export const router = createBrowserRouter([
     {
@@ -13,7 +13,7 @@ export const router = createBrowserRouter([
         element: <SignUpPage />,
     },
     {
-        path: '/log-in',
+        path: '/login',
         element: <LoginPage />,
     },
 ])
