+++ src/schema/login.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const loginSchema = z.object({
+    email: z.string().email({ message: 'Please enter a valid email address' }),
+    password: z.string().min(8, { message: 'Password must be at least 8 characters' }),
+})
