+++ src/schema/signup.ts
@@ -13,8 +13,3 @@ export const signUpSchema = z
         message: "Passwords don't match",
         path: ['confirmPassword'],
     })
-
-export const loginSchema = z.object({
-    email: z.string().email({ message: 'Please enter a valid email address' }),
-    password: z.string().min(8, { message: 'Password must be at least 8 characters' }),
-})
